requirements.txt
django==3.1.4
django-braces==1.14.0
django-extensions==3.1.0
Pillow==8.0.1
pre-commit==2.9.3
